Revert "Change view_model.h to use size_t."
This reverts commit 195d2db161e5cf255d70423f4bfc50cf4efe12ba.

Reason for revert: With this CL in, the browser crashes upon launching on a debug build.

`view_model.h(96) Check failed: index < entries_.size() (0 vs. 0)` 

Original change's description:
> Change view_model.h to use size_t.
>
> Bug: 1292951
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3710687
> Commit-Queue: Peter Kasting <pkasting@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1017319}

Bug: 1292951
No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3722351
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Owners-Override: Luciano Pacheco <lucmult@chromium.org>
Auto-Submit: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1017501}

+++ ash/app_list/app_list_test_api.cc
@@ -98,7 +98,7 @@ views::MenuItemView* GetReorderOptionForNonFolderItemMenu(
 
 ash::AppListItemView* FindFolderItemView(ash::AppsGridView* apps_grid_view) {
   auto* model = apps_grid_view->view_model();
-  for (size_t index = 0; index < model->view_size(); ++index) {
+  for (int index = 0; index < model->view_size(); ++index) {
     ash::AppListItemView* current_view = model->view_at(index);
     if (current_view->is_folder())
       return current_view;
@@ -109,7 +109,7 @@ ash::AppListItemView* FindFolderItemView(ash::AppsGridView* apps_grid_view) {
 
 ash::AppListItemView* FindNonFolderItemView(ash::AppsGridView* apps_grid_view) {
   auto* model = apps_grid_view->view_model();
-  for (size_t index = 0; index < model->view_size(); ++index) {
+  for (int index = 0; index < model->view_size(); ++index) {
     ash::AppListItemView* current_view = model->view_at(index);
     if (!current_view->is_folder())
       return current_view;
@@ -159,7 +159,7 @@ views::MenuItemView* ShowRootMenuAndReturn(
     ui::test::EventGenerator* event_generator) {
   views::MenuItemView* root_menu = nullptr;
 
-  EXPECT_GT(apps_grid_view->view_model()->view_size(), 0u);
+  EXPECT_GT(apps_grid_view->view_model()->view_size(), 0);
 
   switch (menu_type) {
     case AppListTestApi::MenuType::kAppListPageMenu:
@@ -416,7 +416,7 @@ std::u16string AppListTestApi::GetAppListItemViewName(
     const std::string& item_id) {
   views::ViewModelT<AppListItemView>* view_model =
       GetTopLevelAppsGridView()->view_model();
-  for (size_t i = 0; i < view_model->view_size(); ++i) {
+  for (int i = 0; i < view_model->view_size(); ++i) {
     AppListItemView* app_list_item_view = view_model->view_at(i);
     if (app_list_item_view->item()->id() == item_id)
       return app_list_item_view->title()->GetText();
@@ -427,7 +427,7 @@ std::u16string AppListTestApi::GetAppListItemViewName(
 std::vector<std::string> AppListTestApi::GetTopLevelViewIdList() {
   std::vector<std::string> id_list;
   auto* view_model = GetTopLevelAppsGridView()->view_model();
-  for (size_t i = 0; i < view_model->view_size(); ++i) {
+  for (int i = 0; i < view_model->view_size(); ++i) {
     AppListItem* app_list_item = view_model->view_at(i)->item();
     if (app_list_item) {
       id_list.push_back(app_list_item->id());
@@ -634,8 +634,7 @@ ash::AppListItemView* AppListTestApi::FindTopLevelFolderItemView() {
 void AppListTestApi::VerifyTopLevelItemVisibility() {
   auto* view_model = GetTopLevelAppsGridView()->view_model();
   std::vector<std::string> invisible_item_names;
-  for (size_t view_index = 0; view_index < view_model->view_size();
-       ++view_index) {
+  for (int view_index = 0; view_index < view_model->view_size(); ++view_index) {
     auto* item_view = view_model->view_at(view_index);
     if (!item_view->GetVisible())
       invisible_item_names.push_back(item_view->item()->name());
